requirements.txt
Django==5.0.1
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
pytesseract==0.3.10
python-dotenv==1.0.0
